[Coupling] Restructure project by function
