+++ hosting/src/_services/globals/index.ts
