+++ hosting/src/_services/store/shim.d.ts
@@ -1,6 +1,6 @@
 import { Store } from 'vuex'
 
-import { PTVuexState } from '@/store/vuex-api'
+import { PTVuexState } from '~/store/vuex-api'
 
 declare module '@vue/runtime-core' {
   interface ComponentCustomProperties {
