+++ hosting/src/events/Events/Events.ts
@@ -1,8 +1,8 @@
 import { setup, Vue } from 'vue-class-component'
 import { useMeta } from 'vue-meta'
 
-import { EventUi } from '@/models/event'
-import { Route } from '@/router/route-decorator'
+import { EventUi } from '@/events/event'
+import { Route } from '~/router/route-decorator'
 
 @Route({ name: 'Events', path: '/events' })
 export default class Events extends Vue {
