+++ hosting/src/events/EventsCreate/EventsCreate.ts
@@ -1,7 +1,7 @@
 import { setup, Vue } from 'vue-class-component'
 import { useMeta } from 'vue-meta'
 
-import { Route } from '@/router/route-decorator'
+import { Route } from '~/router/route-decorator'
 
 @Route({ name: 'EventsCreate', path: '/events/create' })
 export default class EventsCreate extends Vue {
