+++ hosting/src/main.ts
@@ -2,16 +2,15 @@ import type { App as VueApp } from 'vue'
 import { createApp } from 'vue'
 import { createMetaManager } from 'vue-meta'
 
-// import { firestorePlugin as VueFire } from 'vuefire' // TODO: VueFire for Vue 3
-import router from '@/router'
-import store from '@/store'
-import { ACTIONS } from '@/store/vuex-api'
 import { App } from '@/views'
 import { auth } from '~/firebase-initialized'
-import { globals } from '~/vue/globals'
-import { mixins } from '~/vue/mixins'
+import { globals } from '~/globals'
+import { mixins } from '~/mixins'
+import router from '~/router'
+import store from '~/store'
+import { ACTIONS } from '~/store/vuex-api'
 
-import '@/styles/index.css'
+import '@/main.css'
 
 // Setup Vue
 const create = async (): Promise<VueApp<Element>> => {
@@ -19,7 +18,6 @@ const create = async (): Promise<VueApp<Element>> => {
     .use(router)
     .use(store)
     .use(createMetaManager())
-    // .use(VueFire) // TODO: VueFire for Vue 3
 
   Object.entries(globals).forEach(([key, value]) => app.provide(key, value))
   mixins.forEach(mixin => app.mixin(mixin))
