+++ hosting/src/views/Global/Debug.vue
@@ -71,7 +71,7 @@ import { useMeta } from 'vue-meta'
 import { User } from '@phits-tech/common/dist/dao-firestore/model-types'
 import { DeepRequiredWithId } from '@phits-tech/common/dist/utils/types/general'
 
-import store from '@/store'
+import store from '~/vue/store'
 
 export default class Debug extends Vue {
   meta = setup(() => useMeta({ title: 'Debug' }))
