+++ hosting/src/views/Global/NotFound.vue
@@ -3,7 +3,7 @@ import { Vue } from 'vue-class-component'
 
 import { capitalize } from '@phits-tech/common/dist/utils/string-cases/capitalize'
 
-import { Route } from '@/router/route-decorator'
+import { Route } from '~/router/route-decorator'
 
 @Route({ name: 'NotFound', path: '/:catchAllSegments(.*)', priority: -1000 })
 export default class NotFound extends Vue.with(class {
