+++ hosting/src/views/Profile/Profile.ts
@@ -3,7 +3,7 @@ import { Vue } from 'vue-class-component'
 import { useMeta } from 'vue-meta'
 import { useRoute } from 'vue-router'
 
-import { Route } from '@/router/route-decorator'
+import { Route } from '~/router/route-decorator'
 
 @Route({ name: 'Profile', path: '/profile/:slug' })
 export default class Profile extends Vue {
