+++ hosting/src/views/Spaces/Spaces.ts
@@ -1,7 +1,7 @@
 import { setup, Vue } from 'vue-class-component'
 import { useMeta } from 'vue-meta'
 
-import { Route } from '@/router/route-decorator'
+import { Route } from '~/router/route-decorator'
 
 @Route({ name: 'Spaces', path: '/spaces' })
 export default class Spaces extends Vue {
